fix styling and add correct logic for stale_time and gc_time
